UI: refactored ui codebase into separate files
