+++ include/PluginEditor.hpp
@@ -1,19 +1,15 @@
 #pragma once
 
 #include "PluginProcessor.hpp"
-#include <juce_gui_basics/juce_gui_basics.h>
+#include "CustomLookAndFeel.hpp"      //<-- NEW
+#include "WaveformVisualizer.hpp"     //<-- NEW
 
-class WaveformVisualizer;
-class CantinaLookAndFeel;
-
-//==============================================================================
 class CantinaComposerAudioProcessorEditor  : public juce::AudioProcessorEditor
 {
 public:
     explicit CantinaComposerAudioProcessorEditor (CantinaComposerAudioProcessor&);
     ~CantinaComposerAudioProcessorEditor() override;
 
-    //==============================================================================
     void paint (juce::Graphics&) override;
     void resized() override;
 
@@ -25,17 +21,13 @@ private:
     CantinaComposerAudioProcessor& audioProcessor;
     
     std::unique_ptr<WaveformVisualizer> waveformVisualizer;
-    std::unique_ptr<CantinaLookAndFeel> lookAndFeel;
+    std::unique_ptr<CustomLookAndFeel> lookAndFeel;
 
-    juce::ComboBox presetMenu;
-    juce::ComboBox waveMenu;
-    
+    juce::ComboBox presetMenu, waveMenu;
     juce::Slider attackSlider, decaySlider, sustainSlider, releaseSlider;
-    juce::Label presetLabel, waveLabel, adsrLabel;
-    juce::Label attackLabel, decayLabel, sustainLabel, releaseLabel;
+    juce::Label presetLabel, waveLabel, adsrLabel, attackLabel, decayLabel, sustainLabel, releaseLabel;
 
-    std::unique_ptr<ComboBoxAttachment> presetAttachment;
-    std::unique_ptr<ComboBoxAttachment> waveAttachment;
+    std::unique_ptr<ComboBoxAttachment> presetAttachment, waveAttachment;
     std::unique_ptr<SliderAttachment> attackAttachment, decayAttachment, sustainAttachment, releaseAttachment;
 
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR (CantinaComposerAudioProcessorEditor)
