+++ include/UI/CustomLookAndFeel.hpp
@@ -1,0 +1,11 @@
+#pragma once
+#include <juce_gui_basics/juce_gui_basics.h>
+
+class CustomLookAndFeel : public juce::LookAndFeel_V4
+{
+public:
+    CustomLookAndFeel();
+
+    void drawRotarySlider(juce::Graphics& g, int x, int y, int width, int height, float sliderPos,
+                          const float rotaryStartAngle, const float rotaryEndAngle, juce::Slider& slider) override;
+};
